move moleculeencoder
